move and implement views model to view descriptors model
