+++ src/vs/workbench/browser/parts/views/viewsViewlet.ts
@@ -6,7 +6,7 @@
 import { IAction } from 'vs/base/common/actions';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
-import { IViewDescriptor, IViewDescriptorService } from 'vs/workbench/common/views';
+import { IViewDescriptor, IViewDescriptorService, IAddedViewDescriptorRef } from 'vs/workbench/common/views';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
@@ -16,7 +16,6 @@ import { ViewPaneContainer, ViewPane, IViewPaneOptions } from 'vs/workbench/brow
 import { Event } from 'vs/base/common/event';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/layoutService';
-import { IAddedViewDescriptorRef } from 'vs/workbench/browser/parts/views/views';
 
 export interface IViewletViewOptions extends IViewPaneOptions {
 }
@@ -41,13 +40,13 @@ export abstract class FilterViewPaneContainer extends ViewPaneContainer {
 		@IViewDescriptorService viewDescriptorService: IViewDescriptorService
 	) {
 
-		super(viewletId, `${viewletId}.state`, { mergeViewWithContainerWhenSingleView: false }, instantiationService, configurationService, layoutService, contextMenuService, telemetryService, extensionService, themeService, storageService, contextService, viewDescriptorService);
+		super(viewletId, { mergeViewWithContainerWhenSingleView: false }, instantiationService, configurationService, layoutService, contextMenuService, telemetryService, extensionService, themeService, storageService, contextService, viewDescriptorService);
 		this._register(onDidChangeFilterValue(newFilterValue => {
 			this.filterValue = newFilterValue;
 			this.onFilterChanged(newFilterValue);
 		}));
 
-		this._register(this.viewsModel.onDidChangeActiveViews((viewDescriptors) => {
+		this._register(this.viewsDescriptors.onDidChangeActiveViews((viewDescriptors) => {
 			this.updateAllViews(viewDescriptors);
 		}));
 	}
@@ -63,7 +62,7 @@ export abstract class FilterViewPaneContainer extends ViewPaneContainer {
 			}
 			this.allViews.get(filterOnValue)!.set(descriptor.id, descriptor);
 			if (this.filterValue && !this.filterValue.includes(filterOnValue)) {
-				this.viewsModel.setVisible(descriptor.id, false);
+				this.viewsDescriptors.setVisible(descriptor.id, false);
 			}
 		});
 	}
@@ -76,17 +75,17 @@ export abstract class FilterViewPaneContainer extends ViewPaneContainer {
 
 	private onFilterChanged(newFilterValue: string[]) {
 		if (this.allViews.size === 0) {
-			this.updateAllViews(this.viewsModel.viewDescriptors);
+			this.updateAllViews(this.viewsDescriptors.activeViewDescriptors);
 		}
-		this.getViewsNotForTarget(newFilterValue).forEach(item => this.viewsModel.setVisible(item.id, false));
-		this.getViewsForTarget(newFilterValue).forEach(item => this.viewsModel.setVisible(item.id, true));
+		this.getViewsNotForTarget(newFilterValue).forEach(item => this.viewsDescriptors.setVisible(item.id, false));
+		this.getViewsForTarget(newFilterValue).forEach(item => this.viewsDescriptors.setVisible(item.id, true));
 	}
 
 	getContextMenuActions(): IAction[] {
 		const result: IAction[] = Array.from(this.constantViewDescriptors.values()).map(viewDescriptor => (<IAction>{
 			id: `${viewDescriptor.id}.toggleVisibility`,
 			label: viewDescriptor.name,
-			checked: this.viewsModel.isVisible(viewDescriptor.id),
+			checked: this.viewsDescriptors.isVisible(viewDescriptor.id),
 			enabled: viewDescriptor.canToggleVisibility,
 			run: () => this.toggleViewVisibility(viewDescriptor.id)
 		}));
@@ -134,7 +133,7 @@ export abstract class FilterViewPaneContainer extends ViewPaneContainer {
 		}
 		// Check that allViews is ready
 		if (this.allViews.size === 0) {
-			this.updateAllViews(this.viewsModel.viewDescriptors);
+			this.updateAllViews(this.viewsDescriptors.activeViewDescriptors);
 		}
 		return panes;
 	}
