+++ src/vs/workbench/contrib/scm/browser/scmViewlet.ts
@@ -24,12 +24,11 @@ import { INotificationService } from 'vs/platform/notification/common/notificati
 import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/layoutService';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
-import { IViewsRegistry, Extensions, IViewDescriptorService, IViewDescriptor } from 'vs/workbench/common/views';
+import { IViewsRegistry, Extensions, IViewDescriptorService, IViewDescriptor, IAddedViewDescriptorRef, IViewDescriptorRef } from 'vs/workbench/common/views';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { RepositoryPane, RepositoryViewDescriptor } from 'vs/workbench/contrib/scm/browser/repositoryPane';
 import { MainPaneDescriptor, MainPane, IViewModel } from 'vs/workbench/contrib/scm/browser/mainPane';
 import { ViewPaneContainer, IViewPaneOptions, ViewPane } from 'vs/workbench/browser/parts/views/viewPaneContainer';
-import type { IAddedViewDescriptorRef, IViewDescriptorRef } from 'vs/workbench/browser/parts/views/views';
 import { debounce } from 'vs/base/common/decorators';
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
@@ -80,8 +79,6 @@ export class EmptyPaneDescriptor implements IViewDescriptor {
 
 export class SCMViewPaneContainer extends ViewPaneContainer implements IViewModel {
 
-	private static readonly STATE_KEY = 'workbench.scm.views.state';
-
 	private menus: SCMMenus;
 	private _repositories: ISCMRepository[] = [];
 
@@ -104,7 +101,7 @@ export class SCMViewPaneContainer extends ViewPaneContainer implements IViewMode
 	}
 
 	get onDidChangeVisibleRepositories(): Event<ISCMRepository[]> {
-		const modificationEvent = Event.debounce(Event.any(this.viewsModel.onDidAdd, this.viewsModel.onDidRemove), () => null, 0);
+		const modificationEvent = Event.debounce(Event.any(this.viewsDescriptors.onDidAdd, this.viewsDescriptors.onDidRemove), () => null, 0);
 		return Event.map(modificationEvent, () => this.visibleRepositories);
 	}
 
@@ -126,7 +123,7 @@ export class SCMViewPaneContainer extends ViewPaneContainer implements IViewMode
 		@IContextKeyService contextKeyService: IContextKeyService,
 		@IViewDescriptorService viewDescriptorService: IViewDescriptorService
 	) {
-		super(VIEWLET_ID, SCMViewPaneContainer.STATE_KEY, { mergeViewWithContainerWhenSingleView: true }, instantiationService, configurationService, layoutService, contextMenuService, telemetryService, extensionService, themeService, storageService, contextService, viewDescriptorService);
+		super(VIEWLET_ID, { mergeViewWithContainerWhenSingleView: true }, instantiationService, configurationService, layoutService, contextMenuService, telemetryService, extensionService, themeService, storageService, contextService, viewDescriptorService);
 
 		this.menus = instantiationService.createInstance(SCMMenus, undefined);
 		this._register(this.menus.onDidChangeTitle(this.updateTitleArea, this));
@@ -143,14 +140,14 @@ export class SCMViewPaneContainer extends ViewPaneContainer implements IViewMode
 
 		this._register(configurationService.onDidChangeConfiguration(e => {
 			if (e.affectsConfiguration('scm.alwaysShowProviders') && configurationService.getValue<boolean>('scm.alwaysShowProviders')) {
-				this.viewsModel.setVisible(MainPane.ID, true);
+				this.viewsDescriptors.setVisible(MainPane.ID, true);
 			}
 		}));
 
 		this.repositoryCountKey = contextKeyService.createKey('scm.providerCount', 0);
 
-		this._register(this.viewsModel.onDidAdd(this.onDidShowView, this));
-		this._register(this.viewsModel.onDidRemove(this.onDidHideView, this));
+		this._register(this.viewsDescriptors.onDidAdd(this.onDidShowView, this));
+		this._register(this.viewsDescriptors.onDidRemove(this.onDidHideView, this));
 	}
 
 	create(parent: HTMLElement): void {
@@ -217,8 +214,8 @@ export class SCMViewPaneContainer extends ViewPaneContainer implements IViewMode
 
 	@debounce(0)
 	private afterOnDidHideView(): void {
-		if (this.repositoryCountKey.get()! > 0 && this.viewDescriptors.every(d => !this.viewsModel.isVisible(d.id))) {
-			this.viewsModel.setVisible(this.viewDescriptors[0].id, true);
+		if (this.repositoryCountKey.get()! > 0 && this.viewDescriptors.every(d => !this.viewsDescriptors.isVisible(d.id))) {
+			this.viewsDescriptors.setVisible(this.viewDescriptors[0].id, true);
 		}
 	}
 
@@ -285,9 +282,9 @@ export class SCMViewPaneContainer extends ViewPaneContainer implements IViewMode
 	}
 
 	setVisibleRepositories(repositories: ISCMRepository[]): void {
-		const visibleViewDescriptors = this.viewsModel.visibleViewDescriptors;
+		const visibleViewDescriptors = this.viewsDescriptors.visibleViewDescriptors;
 
-		const toSetVisible = this.viewsModel.viewDescriptors
+		const toSetVisible = this.viewsDescriptors.activeViewDescriptors
 			.filter((d): d is RepositoryViewDescriptor => d instanceof RepositoryViewDescriptor && repositories.indexOf(d.repository) > -1 && visibleViewDescriptors.indexOf(d) === -1);
 
 		const toSetInvisible = visibleViewDescriptors
@@ -305,11 +302,11 @@ export class SCMViewPaneContainer extends ViewPaneContainer implements IViewMode
 				}
 			}
 
-			this.viewsModel.setVisible(viewDescriptor.id, false);
+			this.viewsDescriptors.setVisible(viewDescriptor.id, false);
 		}
 
 		for (const viewDescriptor of toSetVisible) {
-			this.viewsModel.setVisible(viewDescriptor.id, true, size);
+			this.viewsDescriptors.setVisible(viewDescriptor.id, true, size);
 		}
 	}
 }
